build

+++ cluster.c
@@ -8,7 +8,7 @@ void read_row(int i, int n, FILE* input, struct _spmat* A) {
 	int* row;
 	int k, j, cur;
 
-	row = (int*)malloc(n * sizeof(int));
+	row = (double*)malloc(n * sizeof(double));
 	assert(row!=NULL);				/* TODO: error module*/
 
 	for (j = 0; j < n; j++) {
